pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yt-dlp-playlists-downloader"
version = "0.1.0"
description = "Download playlist tracks with yt-dlp, apply metadata from TOML, and optionally use a PyQt6 GUI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPLv3" }
authors = [{ name = "crowdedmovie" }]
dependencies = [
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "mutagen>=1.47.0",
    "PyQt6>=6.7,<7",
    "qtpy>=2.4,<3",
    "qt-themes>=0.4.0",
    "qt-material-icons>=0.5.0",
]

[project.scripts]
yt-dlp-playlists-downloader = "yt_dlp_playlists_downloader.cli:main"
yt-dlp-playlists-downloader-gui = "yt_dlp_playlists_downloader.gui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
yt_dlp_playlists_downloader = ["gui/ui/*.ui", "gui/assets/*"]
